+++ src/test/java/ibm/jceplus/junit/openjceplus/TestAll.java
@@ -64,8 +64,6 @@ import org.junit.platform.suite.api.Suite;
     TestHmacSHA224InteropSunJCE.class,
     TestHmacSHA256.class,
     TestHmacSHA256InteropSunJCE.class,
-    TestHmacSHA3_384.class,
-    TestHmacSHA3_512.class,
     TestHmacSHA384.class,
     TestHmacSHA384InteropSunJCE.class,
     TestHmacSHA512.class,
