+++ src/test/java/ibm/jceplus/junit/tests/TestHmacSHA3_384.java
@@ -1,21 +1,37 @@
 /*
- * Copyright IBM Corp. 2023, 2024
+ * Copyright IBM Corp. 2023, 2026
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms provided by IBM in the LICENSE file that accompanied
  * this code, including the "Classpath" Exception described therein.
  */
 
-package ibm.jceplus.junit.base;
+package ibm.jceplus.junit.tests;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.params.Parameter;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.MethodSource;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
+@Tag(Tags.OPENJCEPLUS_NAME)
+@Tag(Tags.OPENJCEPLUS_FIPS_NAME)
+@Tag(Tags.OPENJCEPLUS_MULTITHREAD_NAME)
+@Tag(Tags.OPENJCEPLUS_FIPS_MULTITHREAD_NAME)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ParameterizedClass
+@MethodSource("ibm.jceplus.junit.tests.TestArguments#getEnabledProviders")
+public class TestHmacSHA3_384 extends BaseTest {
+
+    @Parameter(0)
+    TestProvider provider;
 
     // test vectors fromhttps://csrc.nist.gov/projects/cryptographic-standards-and-guidelines/example-values#aMsgAuth
 
@@ -40,42 +56,45 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
             NISTKEY_172[i] = (byte) i;
         }
         /*
-        
-        t("HmacSHA3-512", STR_NIST1, "4efd629d6c71bf86162658f29943b1c308ce27cdfa6db0d9c3ce81763f9cbce5f7ebe9868031db1a8f8eb7b6b95e5c5e3f657a8996c86a2f6527e307f0213196", Arrays.copyOf(NISTKEY_172, 64)),
-        t("HmacSHA3-512", STR_NIST2, "544e257ea2a3e5ea19a590e6a24b724ce6327757723fe2751b75bf007d80f6b360744bf1b7a88ea585f9765b47911976d3191cf83c039f5ffab0d29cc9d9b6da", Arrays.copyOf(NISTKEY_172, 72)),
-        t("HmacSHA3-512", STR_NIST3, "5f464f5e5b7848e3885e49b2c385f0694985d0e38966242dc4a5fe3fea4b37d46b65ceced5dcf59438dd840bab22269f0ba7febdb9fcf74602a35666b2a32915", Arrays.copyOf(NISTKEY_172, 136)),
-        
+        t("HmacSHA3-384", STR_NIST1, "d588a3c51f3f2d906e8298c1199aa8ff6296218127f6b38a90b6afe2c5617725bc99987f79b22a557b6520db710b7f42", Arrays.copyOf(NISTKEY_172, 48)),
+        t("HmacSHA3-384", STR_NIST2, "a27d24b592e8c8cbf6d4ce6fc5bf62d8fc98bf2d486640d9eb8099e24047837f5f3bffbe92dcce90b4ed5b1e7e44fa90", Arrays.copyOf(NISTKEY_172, 104)),
+        t("HmacSHA3-384", STR_NIST3, "e5ae4c739f455279368ebf36d4f5354c95aa184c899d3870e460ebc288ef1f9470053f73f7c6da2a71bcaec38ce7d6ac", Arrays.copyOf(NISTKEY_172, 152)),
         */
 
-        key_1 = Arrays.copyOf(NISTKEY_172, 64);
+        key_1 = Arrays.copyOf(NISTKEY_172, 48);
 
         data1 = "Sample message for keylen<blocklen";
         data_1 = data1.getBytes(StandardCharsets.UTF_8);
 
         digest_1 = BaseUtils.hexStringToByteArray(
-                "4efd629d6c71bf86162658f29943b1c308ce27cdfa6db0d9c3ce81763f9cbce5f7ebe9868031db1a8f8eb7b6b95e5c5e3f657a8996c86a2f6527e307f0213196");
+                "d588a3c51f3f2d906e8298c1199aa8ff6296218127f6b38a90b6afe2c5617725bc99987f79b22a557b6520db710b7f42");
 
-        key_2 = Arrays.copyOf(NISTKEY_172, 72);
+        key_2 = Arrays.copyOf(NISTKEY_172, 104);
 
         data2 = "Sample message for keylen=blocklen";
         data_2 = data2.getBytes(StandardCharsets.UTF_8);
 
         digest_2 = BaseUtils.hexStringToByteArray(
-                "544e257ea2a3e5ea19a590e6a24b724ce6327757723fe2751b75bf007d80f6b360744bf1b7a88ea585f9765b47911976d3191cf83c039f5ffab0d29cc9d9b6da");
+                "a27d24b592e8c8cbf6d4ce6fc5bf62d8fc98bf2d486640d9eb8099e24047837f5f3bffbe92dcce90b4ed5b1e7e44fa90");
 
-        key_3 = Arrays.copyOf(NISTKEY_172, 136);
+        key_3 = Arrays.copyOf(NISTKEY_172, 152);
 
         data3 = "Sample message for keylen>blocklen";
         data_3 = data3.getBytes(StandardCharsets.UTF_8);
 
         digest_3 = BaseUtils.hexStringToByteArray(
-                "5f464f5e5b7848e3885e49b2c385f0694985d0e38966242dc4a5fe3fea4b37d46b65ceced5dcf59438dd840bab22269f0ba7febdb9fcf74602a35666b2a32915");
+                "e5ae4c739f455279368ebf36d4f5354c95aa184c899d3870e460ebc288ef1f9470053f73f7c6da2a71bcaec38ce7d6ac");
+    }
+
+    @BeforeEach
+    public void setUp() throws Exception {
+        setAndInsertProvider(provider);
     }
 
     @Test
-    public void testHmacSHA3_512_key1() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
-        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-512");
+    public void testHmacSHA3_384_key1() throws Exception {
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
+        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-384");
         mac.init(key);
         mac.update(data_1);
         byte[] digest = mac.doFinal();
@@ -84,9 +103,9 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
     }
 
     @Test
-    public void testHmacSHA3_512_key2() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
-        SecretKeySpec key = new SecretKeySpec(key_2, "HmacSHA3-512");
+    public void testHmacSHA3_384_key2() throws Exception {
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
+        SecretKeySpec key = new SecretKeySpec(key_2, "HmacSHA3-384");
         mac.init(key);
         mac.update(data_2);
         byte[] digest = mac.doFinal();
@@ -95,9 +114,9 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
     }
 
     @Test
-    public void testHmacSHA3_512_key3() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
-        SecretKeySpec key = new SecretKeySpec(key_3, "HmacSHA3-512");
+    public void testHmacSHA3_384_key3() throws Exception {
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
+        SecretKeySpec key = new SecretKeySpec(key_3, "HmacSHA3-384");
         mac.init(key);
         mac.update(data_3);
         byte[] digest = mac.doFinal();
@@ -107,8 +126,8 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
 
     @Test
     public void test_reset() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
-        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-512");
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
+        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-384");
         mac.init(key);
         mac.update(data_1);
         mac.reset();
@@ -120,8 +139,8 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
 
     @Test
     public void test_reuse() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
-        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-512");
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
+        SecretKeySpec key = new SecretKeySpec(key_1, "HmacSHA3-384");
         mac.init(key);
         mac.update(data_1);
         byte[] digest = mac.doFinal();
@@ -136,9 +155,9 @@ public class BaseTestHmacSHA3_512 extends BaseTestJunit5 {
 
     @Test
     public void test_mac_length() throws Exception {
-        Mac mac = Mac.getInstance("HmacSHA3-512", getProviderName());
+        Mac mac = Mac.getInstance("HmacSHA3-384", getProviderName());
         int macLength = mac.getMacLength();
-        boolean isExpectedValue = (macLength == 64);
+        boolean isExpectedValue = (macLength == 48);
         assertTrue(isExpectedValue, "Unexpected mac length");
     }
 }
